Move replacePromBuiltinVariables out of `utils` to work around a Jest issue..
