+++ ui/prometheus-plugin/src/components/PromQLEditor.tsx
@@ -21,7 +21,7 @@ import { ErrorAlert } from '@perses-dev/components';
 import CloseIcon from 'mdi-material-ui/Close';
 import { useReplaceVariablesInString } from '@perses-dev/plugin-system';
 import { PrometheusDatasourceSelector } from '../model';
-import { replacePromBuiltinVariables } from '../utils';
+import { replacePromBuiltinVariables } from '../plugins/prometheus-time-series-query/replace-prom-builtin-variables';
 import { useParseQuery } from './parse';
 import TreeNode from './TreeNode';
