+++ ui/tempo-plugin/package.json
@@ -23,8 +23,8 @@
         "build:types": "tsc --project tsconfig.build.json",
         "type-check": "tsc --noEmit",
         "start": "concurrently -P \"npm:build:* -- {*}\" -- --watch",
-        "test": "TZ=UTC jest",
-        "test:watch": "TZ=UTC jest --watch",
+        "test": "cross-env TZ=UTC jest",
+        "test:watch": "cross-env TZ=UTC jest --watch",
         "lint": "eslint src --ext .ts,.tsx",
         "lint:fix": "eslint --fix src --ext .ts,.tsx"
     },
